Split outlook.com to demote Exchange for Hotmail
We want IMAP to be the default for Hotmail accounts (including hotmail.de,
live.com, outlook.com, outlook.de etc.).

However, due to the MX setup, Office365 accounts also draw the same config
as for outlook.com. Bug 1185366 fixes this, by allowing 2 separate configs
for Office365 and Hotmail.

But that won't help clients already deployed. In the interim, as a hotfix,
split the config into 3 configs:
* Config for Office365. This requires bug 1185366 fixed.
* Config for @hotmail.com, @hotmail.de, @live.com, @live.de, @outlook.de,
  and all other known domains, but not outlook.com. This will be active
  for all clients immediately and default to IMAP instead of Exchange.
* Config for outlook.com only. Until bug 1185366 is fixed, this will be
  used for @outlook.com (exactly this) and all Office365 users.
  For this reason, it stays as is for now. Later fixes will improve this.