Update messagingengine.com to use OAuth2
Support for Fastmail OAuth authentication was added in Thunderbird 115
(https://bugzilla.mozilla.org/show_bug.cgi?id=1785240).

Most Fastmail users have their domains setup so that autoconfig.${domain}
points to our servers and delivers the correct autoconfig setup so
OAuth just works.

However for users that don't have that setup, Thunderbird eventually
falls back to the MX record lookup for the domain and finally the
ISPDB lookup for messagingengine.com. Unfortunately that entry is
outdated and still lists authentication as `password-cleartext`.

This causes Thunderbird to get confused because it displays a standard
"password" entry box for connecting to the IMAP server, but also
simultaneously initiates the OAuth authentication flow, I presume for
calendar/contacts access.

This updates the ISBDB entry to list OAuth2 as the correct authentication
mechanism for messagingengine.com